requirements.txt
streamlit
fastapi
uvicorn
httpx
python-dotenv
pydantic
